# server/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "mfa_required": {
            "description": "Post the authenticator code to /auth/mfa to finish.",
            "title": "Mfa Required",
            "type": "boolean"
          }
        },
        "required": [
          "mfa_required"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "MfaRequest": {
        "properties": {
          "code": {
            "description": "A six-digit authenticator code, or one recovery code.",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "MfaRequest",
        "type": "object"
      },
...
      "PasswordChange": {
        "properties": {
          "current_password": {
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "minLength": 12,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange",
        "type": "object"
      },
      "PasswordConfirm": {
        "description": "Re-entering the password guards changes to the second factor.",
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "PasswordConfirm",
        "type": "object"
      },
...
      "RecoveryCodes": {
        "properties": {
          "codes": {
            "description": "Shown once. They cannot be retrieved again.",
            "items": {
              "type": "string"
            },
            "title": "Codes",
            "type": "array"
          }
        },
        "required": [
          "codes"
        ],
        "title": "RecoveryCodes",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Find My Tracker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "description": "Step one. With two-factor on, this only issues a short-lived ticket for `/auth/mfa`.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/mfa": {
      "post": {
        "description": "Step two: a six-digit authenticator code, or one recovery code.",
        "operationId": "submit_mfa_api_auth_mfa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Mfa",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password": {
      "post": {
        "description": "Changing the password signs every other browser out. This one stays signed in.",
        "operationId": "change_password_api_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/recovery-codes": {
      "post": {
        "description": "Replace every unused recovery code with a fresh set.",
        "operationId": "regenerate_recovery_codes_api_auth_recovery_codes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Recovery Codes",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/totp/confirm": {
      "post": {
        "operationId": "confirm_totp_api_auth_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Totp",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/totp/disable": {
      "post": {
        "operationId": "disable_totp_api_auth_totp_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Totp",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}